chore: migration to MojoTest
